Hash password in /Users API endpoint
Hashing password in /Users endpoint ensures that new user is not hacking the database. Also ensures that only the user knows their password.

CORS is integrated here as well to control which domains have access to this API. Right now CORS is set to default meaning, that all domains have access. This is not good practice and will be changed later.